Error Distance is Not Absolute; This May Not Work
Need to use the absolute error distance.

The tactic of this branch doesn't really work well.  The issue is that it
takes a long time (500 ms??) to calculate the XY position based on the
current chain lengths.  If the motors continue to move during that time,
then by the time the chain calculation finishes we cause the sled to
quickly jump backwards in its movement.

Any solution that doesn't fix a feedrate seems like a bit of a bodge.