Change Loop Interval to Microseconds; Remove Moveent Loop Timer
Instead of running the movement loop on a timer, run whenever the
PID loop has cleared the last move.  There was an odd instance in
which the movement loop was scheduled to happen just before PID
loop.  In that case, the movement loop would at times fail to
complete (maybe becuase of a serial interrupt).  The movement loop
would continue to oscillate like this.

The new code does have the issue wherein if the movement loop fails
to complete, it will get slowed down because it no longer has the
concept of time.  This is fine, if this is happening a lot then
the loop interval is too high.