Add Soft Limits;  Add Critical Alarm
Add concept of Critical_Alarms which are unrecoverable, such as a
loss of position.  They can technically be recovered by fixing an
errant setting, but no machine motions are allowed while in this
stated.

Check soft limits more frequently, including alarming if we are ever
outside the worksurface such as when a bad setting has been entered.

Soft limits are currently checked at the start of any inverse call.
(this is not a good place, should be in motion). As well as after
any forward call to determine if forward result is on worksurface.