Add $MSLWx= Commands; Remove Old B Commands
Add $MSLW5= or 6= command for setting the encoder position, currently
only used for zeroing out things. 5 is left, 6 is right.  Z is not
an option

Add $MSLW7= or 8= for moving an axis an incremental distance in MM.
7 is left, 8 is right.

These are necessary to fit more in the style of using $ commands.
B commands could be a real axis.  Also necessary so that the commands
can be run even when machine is in an alarm state.