Catch Line Ending on Quick Commands; Remove Watchdog
Need to discard any line endings sent after a quick command
otherwise they get interpreted as a blank line and an OK is
sent.  I am not super thrilled with how I wrote this, but I
think it is reasonably compact and logical.

REMOVE WATCHDOG.  This was also causing an issue when the
machine paused itself waiting for a manual Z change.  If
the buffer was empty, it would send an OK requesting a new
line.  We just need to be careful now to make sure that all
OKs are properly sent.