Don't Write to EEPROM Until Receive Serial Command
It looks like at least on two of my machines, that the arduino gets
powered up immediately on being plugged in and then GC connects after
this point.  As a result, we are writing to the EEPROM before we have a
chance to warn the user.

This is a silly solution.  There is no gaurantee that anything written
to the serial port is done by GC.

The real answer is to get the alarm code working in the firmware and
trip an alarm that has to be cleared by the user using a specific
command.

So this would fix #356, but it still needs work in the form of an alarm
structure.